Add local store and set up Firebase
The store saves information of the session, which is sent to Firebase
after the user goes back to home page; then, the store is reset.

The Firebase config is not commited here. Firebase apiKey isn't secret,
and the user can inspect the app bundle to get it anyway, but still, not
showing it publicly is probably better.