Show modal if user leaves for longer than timeout
I tried to navigate to a new page with ReflectionModal there, but
could not figure out how to pass the elapsedMinutes parameter, nor how
to save it to the store correctly. Using ref, using wrapper function,
and putting state in dependency array all fail to capture the state at
the moment the callback is called.